update: long long type

+++ src/utils/get_time_ms.c
@@ -12,7 +12,7 @@
 
 #include "codexion.h"
 
-long	get_time_ms(void)
+long long	get_time_ms(void)
 {
 	struct timeval	tv;
